pyproject.toml
[project]
name = "feature-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langchain-gigachat>=0.5.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "catboost>=1.2.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []
